Move schema v2 workflows into main package
